Use placeholder for repositoryUrl
The intent of InputAdornment is to add a separate prefix/suffix to the text field, such as a dollar sign or kilograms or icon, and not to format the contents of the text, which is how we've been (mis)using it.
The way to go is to use only `placeholder`. But in using `placeholder`, we need to handle the case with and without the 'http' prefix. I added function getRepositoryUrlPath to do that, and handle a few more likely user inputs, too.