Increase Cypress code coverage 232 (#247)
* Update Cypress and its plugins to latest versions

* Move existing tests to folder `pages`

* Rename spec files with dot notation

This is standard in Cypress land, not that it truly matters

* Add ESLint, with Cypress plugin and settings, specifically to `cypress` directory

* Add IntelliSense to the whole folder rather than at the top of individual spec files. See https://docs.cypress.io/guides/tooling/intelligent-code-completion.html#Reference-type-declarations-via-tsconfig

* Fix lint errors and warnings in `cypress` folder

* Rewrite spec files for consistency and best naming practices

- describe instead of context
- smarter use of before / beforeEach
- unique description per file
- `loads` instead of `loading`

* Rename contributors spec file (typo)

* Add 'noEmit' to tsconfig.json - we only want type checking, not compiling output

* Add a TagCreator spec

- I have edited TagCreator.js liberally to rely on `id` and/or `data-cy` for Cypress testing. This could be organized better. This is just a first attempt and I feel that we should discuss this as a team.

* Add ESLint Cypress script to package.json: `npm run lint:cypress`

- rename instrument/coverage script
- group all 4 Cypress scripts adjacently

* Add a Landing spec

Code coverage for `src/pages/Landing` goes from 7% to 78% with this.

* Add a How To Use spec

Code coverage for `src/pages/HowToUse` goes from 40% to 100% with this

* Add tests to Contributors Page

Testing goes up to 73.5 for the project with this addition

* Add start-server-and-test

This package is used to ensure the server is running before attempting to run Cypress. See https://docs.cypress.io/guides/guides/continuous-integration.html#Boot-your-server

* Add nyc settings file

Purpose is to run coverage to our liking: e.g., exclude serviceWorker.js because it's thoroughly tested

* Add npm scripts for coverage reports

Due to a bug in @cypress/instrument-cra, I have added 3 scripts that end in ":report" to generate the correct coverage report. See https://github.com/cypress-io/instrument-cra/issues/188

* Improve and lint TagCreator and its spec

* Update Cypress version

* Exclude non-user-facing 'guides' from coverage reports

* Complete merge branch 'main'

* Restore data-cy's and id's for Cypress spec

Refactor did not account for these, naturally

* Fix MUI syntax errors on 3 containers

* Use placeholder for repositoryUrl

The intent of InputAdornment is to add a separate prefix/suffix to the text field, such as a dollar sign or kilograms or icon, and not to format the contents of the text, which is how we've been (mis)using it.
The way to go is to use only `placeholder`. But in using `placeholder`, we need to handle the case with and without the 'http' prefix. I added function getRepositoryUrlPath to do that, and handle a few more likely user inputs, too.

Co-authored-by: Brent Hengeveld <jafar.ironclad@gmail.com>