Formatted 'About Page' to all screen sizes
I made an update to a component PictureCard to change the sizing to match the About page figma. It is the only place on the site where the component is used. The GetStarted component does not match the figma but I didn't rework just for this page as it might come up when trying to resize the other pages it's included on. We can address it in a later issue.