Added components AddOrgModal and CountrySelect
Can be tested by adding component <AddOrgModal /> to Paceholder page. This component is the (formatted but non-functioning) input form used to add an organization to the index. The CountrySelect component was taken from the MaterialUI site to add a drop down menu to select a country to the AddOrgModal. The AddOrgModal can be accessed from a button that has not been stylized.