Adds AddOrgModal and CountrySelect components (#463)
Merging as discussed. 

* Adds AddOrganization and CountrySelect components

Adds two components - a form (non-functional) for adding an organization and a component that populates the dropdown 'country' section of the AddOrganization form

* Added components AddOrgModal and CountrySelect

Can be tested by adding component <AddOrgModal /> to Paceholder page. This component is the (formatted but non-functioning) input form used to add an organization to the index. The CountrySelect component was taken from the MaterialUI site to add a drop down menu to select a country to the AddOrgModal. The AddOrgModal can be accessed from a button that has not been stylized.

Co-authored-by: Olivia Chiong <75456721+Olivia-Chiong@users.noreply.github.com>