[NO LONGER SEEKING CHANGES TO ALIGNMENT?] Aligned Dropdowns (#505)
merged as agreed at 4/25 meeting

* Aligned Dropdowns

The dropdowns are aligned for large screens. Radical Collaboration redirects to /adding-projects-to-the-index. After this PR is merged, I will create a new issue so that menu headers are bold if the user is on any of the parents' links.

* Styling of Header/Changes to Cypress

In large displays, aligned dropdowns with headers. In small displays, changed parents to be links. Changed Radical Collaboration landing page to 'How to do it' and adjusted hard code in Cypress to reflect.

* Update NavLink.js

Co-authored-by: Olivia Chiong <75456721+Olivia-Chiong@users.noreply.github.com>