reworked donation page
In order to fix a bug where padding from a grid item was obstructing the ability to click a link, I reworked the Donation page to keep items more contained and ensure that the button was clickable at all display sizes. Small modifications were also made to more accurately represent the figma.